Move slack sync logic to hook
